+++ docs/.storybook/preview.js
@@ -1,34 +1,77 @@
 import '../../src/docs.scss'
 import '../../src/index.scss'
 import '../../src/base/index.scss'
+import '../src/stories/helpers/storybook-styles.scss'
 import renderToHTML from '../src/stories/helpers/code-snippet-html-helper'
 
 const customViewports = {
-  widthSM: {
-    name: 'width-sm',
+  minXS: {
+    name: 'XS (min)',
+    styles: {
+      width: '375px',
+      height: '100%'
+    }
+  },
+  maxXS: {
+    name: 'XS (max)',
+    styles: {
+      width: '543px',
+      height: '100%'
+    }
+  },
+  minSM: {
+    name: 'SM (min)',
     styles: {
       width: '544px',
       height: '100%'
     }
   },
-  widthMD: {
-    name: 'width-md',
+  maxSM: {
+    name: 'SM (max)',
+    styles: {
+      width: '767px',
+      height: '100%'
+    }
+  },
+  minMD: {
+    name: 'MD (min)',
     styles: {
       width: '768px',
       height: '100%'
     }
   },
-  widthLG: {
-    name: 'width-lg',
+  maxMD: {
+    name: 'MD (max)',
+    styles: {
+      width: '1011px',
+      height: '100%'
+    }
+  },
+  minLG: {
+    name: 'LG (min)',
     styles: {
       width: '1012px',
       height: '100%'
     }
   },
-  widthXL: {
-    name: 'width-xl',
+  maxLG: {
+    name: 'LG (max)',
+    styles: {
+      width: '1279px',
+      height: '100%'
+    }
+  },
+  minXL: {
+    name: 'XL (min)',
     styles: {
       width: '1280px',
+      height: '100%'
+    }
+  },
+  medXL: {
+    name: 'XL (med)',
+    styles: {
+      width: '1440px',
       height: '100%'
     }
   }
@@ -47,12 +90,22 @@ export const parameters = {
     expanded: true
   },
 
-  layout: 'fullscreen',
+  layout: 'padded',
   html: {
     root: '#story' // target id for html tab (should be direct parent of <Story /> for easy copy/paste)
   },
   viewport: {viewports: customViewports}
 }
+
+const themes = [
+  'light',
+  'light_colorblind',
+  'light_high_contrast',
+  'dark',
+  'dark_dimmed',
+  'dark_high_contrast',
+  'dark_colorblind'
+]
 
 export const globalTypes = {
   theme: {
@@ -61,7 +114,7 @@ export const globalTypes = {
     defaultValue: 'light',
     toolbar: {
       icon: 'circlehollow',
-      items: ['light', 'light_protanopia', 'dark', 'dark_dimmed', 'dark_high_contrast', 'dark_protanopia', 'all'],
+      items: [...themes, 'all'],
       showName: true
     }
   }
@@ -69,87 +122,21 @@ export const globalTypes = {
 
 export const decorators = [
   (Story, context) => {
-    if (context.globals.theme === 'all') {
-      return (
-        <div class="theme-wrap">
-          <div data-color-mode="light" data-light-theme="light" className="story-wrap" id="story">
-            <Story {...context} />
-          </div>
-
-          <div data-color-mode="light" data-light-theme="light_protanopia" className="story-wrap" id="story">
-            <Story {...context} />
-          </div>
-
-          <div data-color-mode="dark" data-dark-theme="dark" className="story-wrap" id="story">
-            <Story {...context} />
-          </div>
-
-          <div data-color-mode="dark" data-dark-theme="dark_dimmed" className="story-wrap" id="story">
-            <Story {...context} />
-          </div>
-
-          <div data-color-mode="dark" data-dark-theme="dark_high_contrast" className="story-wrap" id="story">
-            <Story {...context} />
-          </div>
-
-          <div data-color-mode="dark" data-dark-theme="dark_protanopia" className="story-wrap" id="story">
-            <Story {...context} />
-          </div>
-        </div>
-      )
-    }
-
-    if (context.globals.theme === 'light') {
-      return (
-        <div data-color-mode="light" data-light-theme="light" className="story-wrap" id="story">
-          <Story {...context} />
-        </div>
-      )
-    }
-
-    if (context.globals.theme === 'light_protanopia') {
-      return (
-        <div data-color-mode="light" data-light-theme="light_protanopia" className="story-wrap" id="story">
-          <Story {...context} />
-        </div>
-      )
-    }
-
-    if (context.globals.theme === 'dark') {
-      return (
-        <div data-color-mode="dark" data-dark-theme="dark" className="story-wrap" id="story">
-          <Story {...context} />
-        </div>
-      )
-    }
-
-    if (context.globals.theme === 'dark_dimmed') {
-      return (
-        <div data-color-mode="dark" data-dark-theme="dark_dimmed" className="story-wrap" id="story">
-          <Story {...context} />
-        </div>
-      )
-    }
-
-    if (context.globals.theme === 'dark_high_contrast') {
-      return (
-        <div data-color-mode="dark" data-dark-theme="dark_high_contrast" className="story-wrap" id="story">
-          <Story {...context} />
-        </div>
-      )
-    }
-
-    if (context.globals.theme === 'dark_protanopia') {
-      return (
-        <div data-color-mode="dark" data-dark-theme="dark_protanopia" className="story-wrap" id="story">
-          <Story {...context} />
-        </div>
-      )
-    }
-
     return (
-      <div className="story-wrap" id="story">
-        <Story {...context} />
+      <div class="theme-wrap">
+        { themes.map((theme) => {
+            if (context.globals.theme === theme || context.globals.theme === 'all') {
+              return <div
+                id="story"
+                className="story-wrap"
+                data-color-mode={theme.startsWith('dark') ? 'dark' : 'light'}
+                data-light-theme={theme.startsWith('light') ? theme : undefined}
+                data-dark-theme={theme.startsWith('dark') ? theme : undefined}
+              >
+                <Story {...context} />
+              </div>
+            }
+          })}
       </div>
     )
   }